# archive/openapi.yaml
openapi: 3.0.0
info:
  title: AnimalTrove
  description: API for an animal-based social media application
  version: 0.2.0
servers:
  - url: 'http://165.22.95.135:3000/'
    description: Main API server
...
paths:
...
  /api/users/like:
    post:
      tags:
        - Users
      summary: Like a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - postID
              properties:
                username:
                  type: string
                postID:
                  type: integer
      responses:
        '201':
          description: Successful like operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/users/unlike:
    post:
      tags:
        - Users
      summary: Unlike a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - postID
              properties:
                username:
                  type: string
                postID:
                  type: integer
      responses:
        '201':
          description: Successful unlike operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/users/getLikeCount:
    post:
      tags:
        - Users
      summary: Get the like count of a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - postID
              properties:
                postID:
                  type: integer
      responses:
        '201':
          description: Successful retrieval of like count
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  likeCount:
                    type: integer
  /api/users/dislike:
    post:
      tags:
        - Users
      summary: Dislike a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - postID
              properties:
                username:
                  type: string
                postID:
                  type: integer
      responses:
        '201':
          description: Successful dislike operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/users/undislike:
    post:
      tags:
        - Users
      summary: Undislike a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - postID
              properties:
                username:
                  type: string
                postID:
                  type: integer
      responses:
        '201':
          description: Successful undislike operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/users/getDislikeCount:
    post:
      tags:
        - Users
      summary: Get the dislike count of a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - postID
              properties:
                postID:
                  type: integer
      responses:
        '201':
          description: Successful retrieval of dislike count
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  dislikeCount:
                    type: integer
...